requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-seed==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==24.4.0
filelock==3.13.3
platformdirs==4.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2024.1
virtualenv==20.25.1
